Refactor `SchemaAwareLogDocumentBuilder` into schema package. (#397)
